Ark-bot_V6/requirements.txt
discord
google-oauth2-tool
gspread
